restructuring and minor fixes
